+++ tests/test_historical_timeseries.py
@@ -11,10 +11,10 @@ class TestHistoricalTimeSeries:
         key = "DUMMY_API_KEY"
         self.client = db.Historical(key=key)
 
-    def test_stream_given_invalid_schema_raises_error(self) -> None:
+    def test_get_range_given_invalid_schema_raises_error(self) -> None:
         # Arrange, Act, Assert
         with pytest.raises(ValueError):
-            self.client.timeseries.stream(
+            self.client.timeseries.get_range(
                 dataset="GLBX.MDP3",
                 symbols="ESH1",
                 schema="ticks",  # <--- invalid
@@ -22,10 +22,10 @@ class TestHistoricalTimeSeries:
                 end="2020-12-28T23:00",
             )
 
-    def test_stream_given_invalid_stype_in_raises_error(self) -> None:
+    def test_get_range_given_invalid_stype_in_raises_error(self) -> None:
         # Arrange, Act, Assert
         with pytest.raises(ValueError):
-            self.client.timeseries.stream(
+            self.client.timeseries.get_range(
                 dataset="GLBX.MDP3",
                 symbols="ESH1",
                 schema="mbo",
@@ -34,10 +34,10 @@ class TestHistoricalTimeSeries:
                 stype_in="zzz",  # <--- invalid
             )
 
-    def test_stream_given_invalid_stype_out_raises_error(self) -> None:
+    def test_get_range_given_invalid_stype_out_raises_error(self) -> None:
         # Arrange, Act, Assert
         with pytest.raises(ValueError):
-            self.client.timeseries.stream(
+            self.client.timeseries.get_range(
                 dataset="GLBX.MDP3",
                 symbols="ESH1",
                 schema="mbo",
@@ -47,12 +47,12 @@ class TestHistoricalTimeSeries:
             )
 
     @pytest.mark.skipif(sys.version_info < (3, 8), reason="incompatible mocking")
-    def test_stream_sends_expected_request(self, mocker: MockerFixture) -> None:
+    def test_get_range_sends_expected_request(self, mocker: MockerFixture) -> None:
         # Arrange
         mocked_get = mocker.patch("requests.get")
 
         # Act
-        self.client.timeseries.stream(
+        self.client.timeseries.get_range(
             dataset="GLBX.MDP3",
             symbols="ES.c.0",
             stype_in="smart",
@@ -65,7 +65,7 @@ class TestHistoricalTimeSeries:
         call = mocked_get.call_args.kwargs
         assert (
             call["url"]
-            == f"https://hist.databento.com/v{db.API_VERSION}/timeseries.stream"
+            == f"https://hist.databento.com/v{db.API_VERSION}/timeseries.get_range"
         )
         assert sorted(call["headers"].keys()) == ["accept", "user-agent"]
         assert call["headers"]["accept"] == "application/json"
@@ -87,7 +87,7 @@ class TestHistoricalTimeSeries:
         assert isinstance(call["auth"], requests.auth.HTTPBasicAuth)
 
     @pytest.mark.skipif(sys.version_info < (3, 8), reason="incompatible mocking")
-    def test_stream_with_limit_sends_expected_request(
+    def test_get_range_with_limit_sends_expected_request(
         self,
         mocker: MockerFixture,
     ) -> None:
@@ -95,7 +95,7 @@ class TestHistoricalTimeSeries:
         mocked_get = mocker.patch("requests.get")
 
         # Act
-        self.client.timeseries.stream(
+        self.client.timeseries.get_range(
             dataset="GLBX.MDP3",
             symbols="ESH1",
             schema="trades",
@@ -108,7 +108,7 @@ class TestHistoricalTimeSeries:
         call = mocked_get.call_args.kwargs
         assert (
             call["url"]
-            == f"https://hist.databento.com/v{db.API_VERSION}/timeseries.stream"
+            == f"https://hist.databento.com/v{db.API_VERSION}/timeseries.get_range"
         )
         assert sorted(call["headers"].keys()) == ["accept", "user-agent"]
         assert call["headers"]["accept"] == "application/json"